[2021/04/09 이재현] 채널 요약 조회 쿼리 변경
1. 이번달, 저번달 구해서 조회
2. 트랜잭션 증감율 반올림 추가

+++ src/main/java/com/brchain/core/channel/dto/ChannelSummaryDto.java
@@ -8,11 +8,11 @@ public class ChannelSummaryDto {
 	private String  channelName;  // 채널 이름
 	private int     channelBlock; // 채널 블럭수
 	private int     channelTx;    // 채널 트렌젝션수
-	private Double    preBlockCnt;  // 채널 블럭수
-	private Double    nowBlockCnt;  // 채널 트렌젝션수
-	private Double    preTxCnt;     // 채널 블럭수
-	private Double    nowTxCnt;     // 채널 트렌젝션수
-	private Double    percent;
+	private Double  preBlockCnt;  // 채널 블럭수
+	private Double  nowBlockCnt;  // 채널 트렌젝션수
+	private Double  preTxCnt;     // 채널 블럭수
+	private Double  nowTxCnt;     // 채널 트렌젝션수
+	private Long    percent;
 	private boolean flag;
 
 	public ChannelSummaryDto(String channelName, int channelBlock, int channelTx, Double preBlockCnt, Double nowBlockCnt, Double preTxCnt, Double nowTxCnt) {
